co-locate optimization with implementation
