+++ eval/src/vespa/eval/tensor/dense/CMakeLists.txt
@@ -8,7 +8,6 @@ vespa_add_library(eval_tensor_dense OBJECT
     dense_tensor_address_combiner.cpp
     dense_tensor_builder.cpp
     dense_tensor_cells_iterator.cpp
-    dense_tensor_function_optimizer.cpp
     dense_tensor_view.cpp
     dense_tensor_reduce.cpp
     mutable_dense_tensor_view.cpp
