+++ eval/src/vespa/eval/tensor/dense/dense_xw_product_function.h
@@ -49,7 +49,8 @@ public:
     bool matrixHasCommonDimensionInnermost() const { return _commonDimensionInnermost; }
 
     eval::InterpretedFunction::Instruction compile_self(Stash &stash) const override;
+    static const eval::TensorFunction &optimize(const eval::TensorFunction &expr, Stash &stash);
 };
 
-}
+} // namespace vespalib::tensor
